+++ src/components/JoinSuccess/JoinSuccess.tsx
@@ -1,6 +1,6 @@
 import { useState, useEffect } from 'react';
 import axios from 'axios';
-import Button from '@/components/button/Button';
+import Button from '@/components/Button/Button';
 import * as S from '@/components/JoinSuccess/JoinSuccess.style';
 import { useNavigate, useLocation } from 'react-router-dom';
 import config from '@/config/config.json';
