+++ src/components/Signup/EmailAuthentication/EmailAuthentication.tsx
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from 'react';
 import * as S from '@/components/Signup/EmailAuthentication/EmailAuthentication.style';
-import Button from '@/components/button/Button';
+import Button from '@/components/Button/Button';
 import Sun from '@/assets/image/onbording/EmailAuthentication/sun.svg';
 import Cloud1 from '@/assets/image/onbording/EmailAuthentication/cloud1.svg';
 import Cloud2 from '@/assets/image/onbording/EmailAuthentication/cloud2.svg';
