+++ src/components/Signup/OAuthsignup/oauthsignup.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react'
 import * as S from '@/components/Signup/OAuthsignup/oauthsignup.style';
-import Button from '@/components/button/Button';
+import Button from '@/components/Button/Button';
 import Cloud1 from '@/assets/image/onbording/oauthsignup/Cloud1.svg';
 import Cloud2 from '@/assets/image/onbording/oauthsignup/Cloud2.svg';
 import Cloud3 from '@/assets/image/onbording/oauthsignup/Cloud3.svg';
